초단기 예보, 실황 추가
